requirements.txt
pyrogram
tgcrypto
python-dotenv
motor
dnspython
aiohttp
